Update to V1.1
Have updated so it works with different models (TS-870 and TS-853A). Added more info on screen. Updated so works with a profile of fan speed changes in an array. Lists HDD temps but does not do anything with the info yet.

Tested on my 853a and appears to work ok